Check that local_dest_file and workspace exist before calling rmtree.

+++ bockbuild/package.py
@@ -158,10 +158,12 @@ class Package:
 				try:
 					checkout (self, source, local_dest_file, workspace)
 				except Exception as e:
-					print 'Deleting ' + local_dest_file + ' cache due to git error'
-					shutil.rmtree (local_dest_file, ignore_errors = True)
-					print 'Deleting ' + workspace + ' cache due to git error'
-					shutil.rmtree (workspace, ignore_errors = True)
+					if os.path.exists(local_dest_file):
+						print 'Deleting ' + local_dest_file + ' cache due to git error'
+						shutil.rmtree(local_dest_file, ignore_errors=True)
+					if os.path.exists(workspace):
+						print 'Deleting ' + workspace + ' cache due to git error'
+						shutil.rmtree(workspace, ignore_errors=True)
 
 					# Explicitly reset the working dir to a known directory which has not been deleted
 					# 'git clone' does not work if you are in a directory which has been deleted
